[REFACTOR] Post `UiState` 위치 변경
