+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/createpost/CreatePostEvent.kt
@@ -1,5 +1,7 @@
 package com.boostcampwm2023.snappoint.presentation.createpost
 
+import com.boostcampwm2023.snappoint.presentation.model.PositionState
+
 sealed class CreatePostEvent {
     data class ShowMessage(val resId: Int): CreatePostEvent()
     data class FindAddress(val index: Int, val position: PositionState) : CreatePostEvent()
