+++ android/app/src/main/java/com/boostcampwm2023/snappoint/presentation/main/MainUiState.kt
@@ -1,8 +1,7 @@
-package com.boostcampwm2023.snappoint.presentation.around
+package com.boostcampwm2023.snappoint.presentation.main
 
 import com.boostcampwm2023.snappoint.presentation.model.PostSummaryState
 
-data class AroundUiState(
+data class MainUiState(
     val posts: List<PostSummaryState> = emptyList()
 )
-
